More Firebase Updates
